feat(ChattingRoomEnterWithProfileResp) : #379 -프로필 입력해서 채팅방 입력DTO생성
-프로필 입력해서 채팅방 입력DTO생성

Related to: #379
