+++ backend/chatting-service/src/main/java/com/chattingservice/api/chatting/dto/request/ChatMessagePageDto.java
@@ -25,6 +25,40 @@ public class ChatMessagePageDto {
     @JsonProperty("meta")
     private PageMetaDto meta;
 
+    @Getter
+    @Builder
+    @AllArgsConstructor
+    @NoArgsConstructor
+    public static class PageMetaDto {
+
+        @JsonProperty("page_number")
+        private int pageNumber;
+
+        @JsonProperty("page_size")
+        private int pageSize;
+
+        @JsonProperty("total_elements")
+        private long totalElements;
+
+        @JsonProperty("total_pages")
+        private int totalPages;
+
+        @JsonProperty("last_flag")
+        private boolean lastFlag;
+
+        static public PageMetaDto from(Page<MessageCollection> chatMessageDtos) {
+            return PageMetaDto.builder()
+                    .pageNumber(chatMessageDtos.getNumber())
+                    .pageSize(chatMessageDtos.getSize())
+                    .totalElements(chatMessageDtos.getTotalElements())
+                    .totalPages(chatMessageDtos.getTotalPages())
+                    .lastFlag(chatMessageDtos.isLast())
+                    .build();
+        }
+
+
+    }
+
     static public ChatMessagePageDto from(Page<MessageCollection> chatMessageDtos) {
         return ChatMessagePageDto.builder()
                 .meta(PageMetaDto.from(chatMessageDtos))
